Vehicle Management System/servo/src/test/java/services/LoginAuthenticatorTest.java
package services;

import static org.junit.Assert.assertTrue;
import static org.junit.jupiter.api.Assertions.assertEquals;

import java.util.Optional;

import org.junit.jupiter.api.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringRunner;

import prime.servo.ServoApplication;
import prime.servo.data.User;
import prime.servo.services.LoginAuthenticator;

@SpringBootTest
@ContextConfiguration(classes = ServoApplication.class)
@RunWith(SpringRunner.class)
class LoginAuthenticatorTest {

	@Autowired
	LoginAuthenticator la;

	@Test
	void login_authenticator_fetch_match() {
		Optional<User> user = la.authenticate("liya", "liya");

		assertTrue(user.isPresent());
		if (user.isPresent()) {
			assertEquals("liya", user.get().getUsername());
		}
	}

}
